- quarantining instances of clientside classes to their own area so that it doesn't crash servers, because I somehow didn't check this - split lots of clientside events from FlatLights.java to ClientEvents, ClientProxy, ClientSetup - ClientUtils for calling on minecraft instances and client player entities - fixed all ClientPlayerEntity calls in various network packets for notifications and clientside updates
