+++ src/main/java/com/uberhelixx/flatlights/network/PacketRisingHeatUpdate.java
@@ -1,6 +1,7 @@
 package com.uberhelixx.flatlights.network;
 
 import com.uberhelixx.flatlights.capability.RisingHeatStateCapability;
+import com.uberhelixx.flatlights.util.ClientUtils;
 import net.minecraft.client.Minecraft;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.LivingEntity;
@@ -32,7 +33,7 @@ public class PacketRisingHeatUpdate {
         if (ctx.get().getDirection().getReceptionSide().isClient()) {
             ctx.get().enqueueWork(() -> {
                 //should be only clientside stuff
-                PlayerEntity player = Minecraft.getInstance().player;
+                PlayerEntity player = ClientUtils.getPlayer();
                 boolean state = msg.state;
                 if(player != null) {
                     Entity entity = player.getEntityWorld().getEntityByID(msg.entityID);
