+++ src/main/java/com/uberhelixx/flatlights/util/ClientUtils.java
@@ -1,0 +1,15 @@
+package com.uberhelixx.flatlights.util;
+
+import net.minecraft.client.Minecraft;
+import net.minecraft.entity.player.PlayerEntity;
+
+public class ClientUtils {
+    
+    public static Minecraft getMinecraft() {
+        return Minecraft.getInstance();
+    }
+    
+    public static PlayerEntity getPlayer() {
+        return getMinecraft().player;
+    }
+}
