+++ src/main/java/com/uberhelixx/flatlights/util/ServerProxy.java
@@ -1,0 +1,20 @@
+package com.uberhelixx.flatlights.util;
+
+import net.minecraftforge.fml.event.lifecycle.FMLCommonSetupEvent;
+
+public class ServerProxy implements IProxy{
+    @Override
+    public void preInit(FMLCommonSetupEvent event) {
+    
+    }
+    
+    @Override
+    public void init(FMLCommonSetupEvent event) {
+    
+    }
+    
+    @Override
+    public void postInit(FMLCommonSetupEvent event) {
+    
+    }
+}
